+++ backend/app/config/config.py
@@ -7,6 +7,7 @@ import os
 import logging
 from dotenv import load_dotenv
 
+
 load_dotenv()
 
 # Токен Телеграм бота
